Created Enemy Spawner Script
Created an early version of the Enemy Spawner.
The basic idea is that when the start button is pressed, it sends the level number to the script, and based on the level it spawns the level number times ten (so level 1 would spawn 10 enemies and level 5 would spawn 50 enemies).
The enemies then spawn from around the spawner prefab and keep spawning at random intervals until the max number of enemies has spawned or the stop button is pressed.