Replay, Restart, and Next Level Scripts
Created a script that reloads the entire scene, so the player can re-select the spire locations. Needs to be tied to a button.
Created a script that replays the level at the same difficulty, intended to be used when the player dies. Needs to be tied to a button.
Created a script that restarts the game at the next highest level, intended to be used when a player successfully completes a level. Needs to be tied to a button.
Changed GamePlayLoop script to have a public level int to keep better track of current level within the game.