CargoShipAI script
Created the script for the Cargo Ship to circle around a random enemy spawner and drop Health and Sloth prefabs when hit. When both prefabs have been gotten, the ship despawns.
Also edited the Bullet script to randomly call the DropHealth() and DropSloth() methods when the Cargo Ship is hit. Still need to implement how to determine which spawner the Cargo Ship will circle around.