Added an additional scene with a spot for an animation
there are issues with github and the project already so something i noticed was that it said that the camera was deleted in the main scene which is something I did not think I even touched. There is also conflicting files with something about the cargo ship AI which i dont understand either.